feat(ai): unify invocation commands and long-running option
